Properly work with optional dependency
